requirements.txt
PyGithub==2.1.1
openai==1.6.1
anthropic==0.8.1
azure-identity==1.15.0
azure-ai-openai==1.0.0b1
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
gitpython==3.1.40
bandit==1.7.6
pylint==3.0.3
radon==6.0.1
safety==2.3.5
semgrep==1.52.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0
google-generativeai==0.3.2
